requirements.txt
imblearn
scikit-learn==1.2.2
matplotlib
seaborn
missingno==0.5.2
regex==2023.6.3
sklearn-features==0.0.2
tqdm==4.65.0
typing-extensions==4.5.0
tzdata==2023.3
pandas== 1.5.3
numpy==1.24.3
matplotlib== 3.7.1 
seaborn== 0.12.2
python== 3.11.3
seaborn==0.12.2
os
streamlit==1.27.1

